backend/requirements.txt
fastapi>=0.110.1
uvicorn[standard]>=0.25.0
motor>=3.3.1
pymongo>=4.5.0
pydantic[email]>=2.6.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
pyjwt>=2.10.1
bcrypt>=4.1.3
passlib>=1.7.4
firebase-admin>=6.5.0
slowapi>=0.1.9
Pillow>=10.0.0
requests>=2.31.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.8
tzdata>=2024.2
